@@ -16,9 +16,12 @@
     "@fontsource/roboto": "^5.0.8",
     "@mui/icons-material": "^5.15.10",
     "@mui/material": "^5.15.10",
+    "ai": "^3.1.30",
     "babel-preset-jest": "^29.6.3",
     "cookie": "^0.6.0",
-    "next": "14.1.0",
+    "dotenv": "^16.4.5",
+    "next": "^14.2.3",
+    "node-fetch": "^2.6.6",
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.51.1",
@@ -37,6 +40,7 @@
     "@types/cookie": "^0.6.0",
     "@types/jest": "^29.5.11",
     "@types/node": "^20.11.25",
+    "@types/node-fetch": "^2.6.11",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18",
     "babel-jest": "^29.7.0",
